Add files via upload
Hello, this is my project. It is an application of what you have learned in the C++ language only. It is a banking system for withdrawing and depositing the necessary amount. It consists of 7 tasks. They are number 1: It displays me the complete list of customer accounts, 2 adds new customers, and 3 deletes unwanted customers.  Among them 4 changes customer data 5 inquires about one customer only 6 consists of 4 tasks No. 1 deposits the required amount 2 withdraws from the customer’s account the required amount 3 which is the list of customers and the total balance in the bank 4 and it returns me to the main menu and the last thing in the main menu  He is exiting the program